refactor moderation commands
